# backend/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IMedicationsData": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"dosage": {
						"type": "number",
						"format": "double"
					},
					"time": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"dosage",
					"time"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IMedicationsData.Exclude_keyofIMedicationsData.id__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"dosage": {
						"type": "number",
						"format": "double"
					},
					"time": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"dosage",
					"time"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IMedicationsData.id_": {
				"$ref": "#/components/schemas/Pick_IMedicationsData.Exclude_keyofIMedicationsData.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Pick_IResetMedicationsData.Exclude_keyofIResetMedicationsData.taken__": {
				"properties": {
					"reminder": {
						"type": "boolean"
					}
				},
				"required": [
					"reminder"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IResetMedicationsData.taken_": {
				"$ref": "#/components/schemas/Pick_IResetMedicationsData.Exclude_keyofIResetMedicationsData.taken__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Pick_IResetMedicationsData.Exclude_keyofIResetMedicationsData.reminder__": {
				"properties": {
					"taken": {
						"type": "boolean"
					}
				},
				"required": [
					"taken"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IResetMedicationsData.reminder_": {
				"$ref": "#/components/schemas/Pick_IResetMedicationsData.Exclude_keyofIResetMedicationsData.reminder__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
...
			"IContactsData": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"relationship": {
						"type": "string"
					},
					"isMainContact": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"name",
					"phone",
					"relationship",
					"isMainContact"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IContactsData.Exclude_keyofIContactsData.id__": {
				"properties": {
					"phone": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"relationship": {
						"type": "string"
					},
					"isMainContact": {
						"type": "boolean"
					}
				},
				"required": [
					"phone",
					"name",
					"relationship",
					"isMainContact"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IContactsData.id_": {
				"$ref": "#/components/schemas/Pick_IContactsData.Exclude_keyofIContactsData.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
...
			"IAppointmentScheduleData": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"doctorId": {
						"type": "string"
					},
					"locationId": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"time": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"id",
					"doctorId",
					"locationId",
					"date",
					"time",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IAppointmentScheduleData.Exclude_keyofIAppointmentScheduleData.id__": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"time": {
						"type": "string"
					},
					"doctorId": {
						"type": "string"
					},
					"locationId": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"time",
					"doctorId",
					"locationId",
					"date",
					"createdAt"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IAppointmentScheduleData.id_": {
				"$ref": "#/components/schemas/Pick_IAppointmentScheduleData.Exclude_keyofIAppointmentScheduleData.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IConfirmScheduleData": {
				"properties": {
					"confirmed": {
						"type": "boolean"
					}
				},
				"required": [
					"confirmed"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "care-idosos",
		"version": "1.0.0",
		"description": "Documentação da API Care Idosos",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Beatriz Ribeiro"
		}
	},
	"paths": {
...
		"/medication": {
			"get": {
				"operationId": "GetMedications",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IMedicationsData"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Medicações"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "AddMedication",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Medicações"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IMedicationsData.id_"
							}
						}
					}
				}
			}
		},
		"/medication/{id}": {
			"get": {
				"operationId": "GetMedicationById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMedicationsData"
								}
							}
						}
					}
				},
				"tags": [
					"Medicações"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "RemoveMedication",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Medicações"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/medication/reminder/{id}": {
			"put": {
				"operationId": "UpdateMedicationReminder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Medicações"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IResetMedicationsData.taken_"
							}
						}
					}
				}
			}
		},
		"/medication/taken/{id}": {
			"put": {
				"operationId": "UpdateMedicationTaken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Medicações"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IResetMedicationsData.reminder_"
							}
						}
					}
				}
			}
		},
		"/medication/reset": {
			"put": {
				"operationId": "ResetMedications",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Medicações"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
		"/contacts": {
			"get": {
				"operationId": "GetEmergencyContacts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IContactsData"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Contatos de Emergência"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "AddEmergencyContact",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Contatos de Emergência"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IContactsData.id_"
							}
						}
					}
				}
			}
		},
		"/contacts/{id}": {
			"get": {
				"operationId": "GetEmergencyContactById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IContactsData"
								}
							}
						}
					}
				},
				"tags": [
					"Contatos de Emergência"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateEmergencyContact",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Contatos de Emergência"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IContactsData.id_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveEmergencyContact",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Contatos de Emergência"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/appointment": {
			"get": {
				"operationId": "GetSchedule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IAppointmentScheduleData"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Agendamento de Consultas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "AddSchedule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Agendamento de Consultas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IAppointmentScheduleData.id_"
							}
						}
					}
				}
			}
		},
		"/appointment/{id}": {
			"get": {
				"operationId": "GetScheduleById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAppointmentScheduleData"
								}
							}
						}
					}
				},
				"tags": [
					"Agendamento de Consultas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateSchedule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Agendamento de Consultas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IAppointmentScheduleData.id_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveSchedule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Agendamento de Consultas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/appointment/confirmed/{id}": {
			"put": {
				"operationId": "ConfirmSchedule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConfirmScheduleData"
								}
							}
						}
					}
				},
				"tags": [
					"Agendamento de Consultas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IConfirmScheduleData"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "https://elderly-care.onrender.com",
			"description": "Produção"
		},
		{
			"url": "https://localhost:3000",
			"description": "Produção"
		}
	]
}